Quest class added for storing the individual quests in the list.
The list has been hard-coded into MainActivity as an array for now, will change this to an ArrayList or similar structure later.

Note: If it's possible, it might be easier to use a struct instead of a class for storing the quests.  I'll look into that later, if I can.