Created three new activities to add additional content:
1) DetailsActivity
2) SettingsActivity
3) RegisterActivity

The current plan is to flush out what I can of each of these activities as they're listed (DetailsActivity first, and so on), then try to work out the parts of the app that connect the different activities together.

I've discovered the usefulness of widgets, and have been using them in the development of this app.